Refactor Jetpack Search Thank You note
